Fix added methods to the wrong file
