+++ crypto-ffi/src/generic/e2ei_context.rs
@@ -1,10 +1,11 @@
-use std::ops::DerefMut;
+use std::{collections::HashMap, ops::DerefMut};
 
-use core_crypto::CryptoError;
+use core_crypto::{prelude::VerifiableGroupInfo, CryptoError, MlsError};
+use tls_codec::Deserialize;
 
 use super::{
-    context::CoreCryptoContext, Ciphersuite, CoreCryptoResult, CrlRegistration, E2eiConversationState, E2eiEnrollment,
-    RotateBundle,
+    context::CoreCryptoContext, Ciphersuite, ClientId, CoreCryptoResult, CrlRegistration, E2eiConversationState,
+    E2eiDumpedPkiEnv, E2eiEnrollment, MlsCredentialType, RotateBundle, WireIdentity,
 };
 
 #[uniffi::export]
@@ -160,5 +161,68 @@ impl CoreCryptoContext {
             .e2ei_conversation_state(&conversation_id)
             .await
             .map(Into::into)?)
+    }
+
+    pub async fn e2ei_dump_pki_env(&self) -> CoreCryptoResult<Option<E2eiDumpedPkiEnv>> {
+        Ok(self.context.e2ei_dump_pki_env().await?.map(Into::into))
+    }
+
+    /// See [core_crypto::mls::MlsCentral::e2ei_is_pki_env_setup]
+    pub async fn e2ei_is_pki_env_setup(&self) -> CoreCryptoResult<bool> {
+        Ok(self.context.e2ei_is_pki_env_setup().await?)
+    }
+
+    /// See [core_crypto::mls::MlsCentral::e2ei_is_enabled]
+    pub async fn e2ei_is_enabled(&self, ciphersuite: Ciphersuite) -> CoreCryptoResult<bool> {
+        let sc = core_crypto::prelude::MlsCiphersuite::from(core_crypto::prelude::CiphersuiteName::from(ciphersuite))
+            .signature_algorithm();
+        Ok(self.context.e2ei_is_enabled(sc).await?)
+    }
+
+    /// See [core_crypto::mls::MlsCentral::get_device_identities]
+    pub async fn get_device_identities(
+        &self,
+        conversation_id: Vec<u8>,
+        device_ids: Vec<ClientId>,
+    ) -> CoreCryptoResult<Vec<WireIdentity>> {
+        let device_ids = device_ids.into_iter().map(|cid| cid.0).collect::<Vec<_>>();
+        Ok(self
+            .context
+            .get_device_identities(&conversation_id, &device_ids[..])
+            .await?
+            .into_iter()
+            .map(Into::into)
+            .collect::<Vec<_>>())
+    }
+
+    /// See [core_crypto::mls::MlsCentral::get_user_identities]
+    pub async fn get_user_identities(
+        &self,
+        conversation_id: Vec<u8>,
+        user_ids: Vec<String>,
+    ) -> CoreCryptoResult<HashMap<String, Vec<WireIdentity>>> {
+        Ok(self
+            .context
+            .get_user_identities(&conversation_id, &user_ids[..])
+            .await?
+            .into_iter()
+            .map(|(k, v)| (k, v.into_iter().map(Into::into).collect()))
+            .collect::<HashMap<String, Vec<WireIdentity>>>())
+    }
+
+    /// See [core_crypto::mls::MlsCentral::get_credential_in_use]
+    pub async fn get_credential_in_use(
+        &self,
+        group_info: Vec<u8>,
+        credential_type: MlsCredentialType,
+    ) -> CoreCryptoResult<E2eiConversationState> {
+        let group_info = VerifiableGroupInfo::tls_deserialize(&mut group_info.as_slice())
+            .map_err(MlsError::from)
+            .map_err(CryptoError::from)?;
+        Ok(self
+            .context
+            .get_credential_in_use(group_info, credential_type.into())
+            .await?
+            .into())
     }
 }
